Move RobotInformation to mbf_utility, as it can be used generaly
