+++ mbf_abstract_nav/include/mbf_abstract_nav/abstract_action_base.hpp
@@ -41,7 +41,7 @@
 
 #include <actionlib/server/action_server.h>
 #include <mbf_abstract_nav/MoveBaseFlexConfig.h>
-#include "mbf_abstract_nav/robot_information.h"
+#include <mbf_utility/robot_information.h>
 
 namespace mbf_abstract_nav{
 
@@ -62,7 +62,7 @@ class AbstractActionBase
 
   AbstractActionBase(
       const std::string& name,
-      const RobotInformation &robot_info,
+      const mbf_utility::RobotInformation &robot_info,
       const RunMethod run_method
   ) : name_(name), robot_info_(robot_info), run_(run_method){}
 
@@ -168,7 +168,7 @@ class AbstractActionBase
 
 protected:
   const std::string &name_;
-  const RobotInformation &robot_info_;
+  const mbf_utility::RobotInformation &robot_info_;
 
   RunMethod run_;
   boost::thread_group threads_;
