+++ mbf_abstract_nav/include/mbf_abstract_nav/controller_action.h
@@ -41,11 +41,13 @@
 #ifndef MBF_ABSTRACT_NAV__CONTROLLER_ACTION_H_
 #define MBF_ABSTRACT_NAV__CONTROLLER_ACTION_H_
 
+#include <actionlib/server/action_server.h>
+
+#include <mbf_msgs/ExePathAction.h>
+#include <mbf_utility/robot_information.h>
+
 #include "mbf_abstract_nav/abstract_action_base.hpp"
 #include "mbf_abstract_nav/abstract_controller_execution.h"
-#include "mbf_abstract_nav/robot_information.h"
-#include <actionlib/server/action_server.h>
-#include <mbf_msgs/ExePathAction.h>
 
 namespace mbf_abstract_nav
 {
@@ -58,7 +60,7 @@ class ControllerAction :
   typedef boost::shared_ptr<ControllerAction> Ptr;
 
   ControllerAction(const std::string &name,
-                   const RobotInformation &robot_info);
+                   const mbf_utility::RobotInformation &robot_info);
 
   /**
    * @brief Start controller action.
