+++ mbf_abstract_nav/include/mbf_abstract_nav/planner_action.h
@@ -41,11 +41,14 @@
 #ifndef MBF_ABSTRACT_NAV__PLANNER_ACTION_H_
 #define MBF_ABSTRACT_NAV__PLANNER_ACTION_H_
 
+#include <actionlib/server/action_server.h>
+
+#include <mbf_msgs/GetPathAction.h>
+#include <mbf_utility/robot_information.h>
+
 #include "mbf_abstract_nav/abstract_action_base.hpp"
 #include "mbf_abstract_nav/abstract_planner_execution.h"
-#include "mbf_abstract_nav/robot_information.h"
-#include <actionlib/server/action_server.h>
-#include <mbf_msgs/GetPathAction.h>
+
 
 namespace mbf_abstract_nav{
 
@@ -57,8 +60,8 @@ class PlannerAction : public AbstractActionBase<mbf_msgs::GetPathAction, Abstrac
   typedef boost::shared_ptr<PlannerAction> Ptr;
 
   PlannerAction(
-      const std::string& name,
-      const RobotInformation &robot_info
+      const std::string &name,
+      const mbf_utility::RobotInformation &robot_info
   );
 
   void run(GoalHandle &goal_handle, AbstractPlannerExecution &execution);
