+++ mbf_utility/CMakeLists.txt
@@ -23,6 +23,7 @@ include_directories(
 
 add_library(${PROJECT_NAME}
    src/navigation_utility.cpp
+   src/robot_information.cpp
 )
 
 add_dependencies(${PROJECT_NAME} ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
