+++ mbf_utility/include/mbf_utility/get_path_exception.h
@@ -6,7 +6,6 @@
 
 namespace mbf_utility{
 struct GetPathException : public std::exception {
-
   GetPathException(unsigned int error_code) : outcome(error_code){}
 
   const char * what () const throw () {
