+++ mbf_utility/include/mbf_utility/robot_information.h
@@ -36,22 +36,22 @@
  *
  */
 
-#ifndef MBF_ABSTRACT_NAV__ROBOT_INFORMATION_H_
-#define MBF_ABSTRACT_NAV__ROBOT_INFORMATION_H_
+#ifndef MBF_UTILITY__ROBOT_INFORMATION_H_
+#define MBF_UTILITY__ROBOT_INFORMATION_H_
 
 #include <boost/shared_ptr.hpp>
 #include <geometry_msgs/PoseStamped.h>
 #include <geometry_msgs/TwistStamped.h>
 #include <ros/duration.h>
 #include <string>
-#include <tf/transform_listener.h>
 
-#include <mbf_utility/types.h>
+#include "mbf_utility/types.h"
 
-namespace mbf_abstract_nav{
+namespace mbf_utility
+{
 
-class RobotInformation{
-
+class RobotInformation
+{
  public:
 
   typedef boost::shared_ptr<RobotInformation> Ptr;
@@ -90,6 +90,6 @@ class RobotInformation{
 
 };
 
-} /* mbf_abstract_nav */
+} /* mbf_utility */
 
-#endif /* MBF_ABSTRACT_NAV__ROBOT_INFORMATION_H_ */
+#endif /* MBF_UTILITY__ROBOT_INFORMATION_H_ */
